+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Upgrade test of the Dogu in Jenkins pipeline (#7)
+- Possibility to set log level for the dogu (#7)
 - Possibility to set memory limit for the dogu (#7)
  
 ## [v3.5.9-1] - 2021-03-22